Task 26: Replacing Generate with Insert
This commit is a simple name change for the commands that will be used
for the CLI tool from `generate` to `insert`
